// types/publish.ts
export interface CurrentAsset {
  tipoActivo: string;
  image?: string;
  nombre: string;
  dni: string;
  activo: string;
  placa: string;
  fechaFabricacion: string;
  categoriaVehiculo: string;
  numeroMotor: string;
  marca: string;
  modelo: string;
  kilometraje: string;
  numeroChasis: string;
  claseVehiculo: string;
  potencia: string;
  tipoCombustible: string;
  numeroSerie: string;
  pesoNeto: string;
  cargaUtil: string;
  pesoBruto: string;
  dimensiones: string;
  color: string;
  carroceria: string;
  ejes: string;
  licenciaA3: Date;
  licenciaA4: Date;
  certificadoSalud: Date;
  recordConductor: Date;
  iqbfConductor: Date;
  manejoDefensivo: Date;
  seguroVidaLey: Date;
  sctrSalud: Date;
  sctrPension: Date;
  habilitacionVehicular: Date;
  resolucionMaterialesPeligrosos: Date;
  inspeccionTecnica: Date;
  planContingencia: Date;
  RDHabilitacion: Date;
  partidaRegistral: Date;
  sunatIQBF: Date;
  soat: Date;
  polizaResponsabilidadCivil: Date;
  NombreArea: string;
  FichaRUC: Date;
  SeguroCarga: Date;
  fechaPostFormato: string;
  fechaPostISO: string;
  createdAt: Date;
  LastEventPosted: Date;
  photoServiceURL: string;
  emailPerfil: string;
  nombrePerfil: string;
  companyName: string;
  idFirebaseAsset: string;
}